+++ src/eth/eip_1559_transaction.rs
@@ -4,6 +4,9 @@
 //! This module contains the logic for EIP-1559 transaction types.
 //! Constants are taken from [FIP-0091](https://github.com/filecoin-project/FIPs/blob/020bcb412ee20a2879b4a710337959c51b938d3b/FIPS/fip-0091.md).
 
+use crate::message::SignedMessage;
+use crate::shim::address::Address;
+use crate::shim::crypto::SignatureType::Delegated;
 use anyhow::ensure;
 use derive_builder::Builder;
 use num::BigInt;
@@ -33,9 +36,39 @@ pub struct EthEip1559TxArgs {
 }
 
 impl EthEip1559TxArgs {
+    /// Returns EIP-1559 transaction signature
+    pub fn signature(&self) -> anyhow::Result<Signature> {
+        // Convert r, s, v to byte arrays
+        let r_bytes = self.r.to_bytes_be().1;
+        let s_bytes = self.s.to_bytes_be().1;
+        let v_bytes = self.v.to_bytes_be().1;
+
+        // Pad r and s to 32 bytes
+        let mut sig = pad_leading_zeros(r_bytes, 32);
+        sig.extend(pad_leading_zeros(s_bytes, 32));
+
+        if v_bytes.is_empty() {
+            sig.push(0);
+        } else {
+            sig.push(*v_bytes.first().expect("failed to get first byte of V"));
+        }
+
+        ensure!(sig.len() == EIP_1559_SIG_LEN, "invalid signature length");
+
+        Ok(Signature {
+            sig_type: Delegated,
+            bytes: sig,
+        })
+    }
+
+    /// Returns a verifiable signature for EIP-1559 transaction
+    pub fn to_verifiable_signature(&self, sig: Vec<u8>) -> anyhow::Result<Vec<u8>> {
+        Ok(sig)
+    }
+
     pub fn with_signature(mut self, signature: &Signature) -> anyhow::Result<Self> {
         ensure!(
-            signature.signature_type() == crate::shim::crypto::SignatureType::Delegated,
+            signature.signature_type() == Delegated,
             "Signature is not delegated type, is {}",
             signature.signature_type()
         );
@@ -62,9 +95,10 @@ impl EthEip1559TxArgs {
         Ok(self)
     }
 
-    pub fn rlp_signed_message(&self) -> anyhow::Result<Vec<u8>> {
+    /// Returns an RLP stream representing the EIP-1559 transaction message
+    fn message_rlp_stream(&self) -> anyhow::Result<rlp::RlpStream> {
         // https://github.com/filecoin-project/lotus/blob/v1.27.1/chain/types/ethtypes/eth_1559_transactions.go#L72
-        let prefix = [EIP_1559_TX_TYPE as u8].as_slice();
+        let prefix = [EIP_1559_TX_TYPE].as_slice();
         let access_list: &[u8] = &[];
         let mut stream = rlp::RlpStream::new_with_buffer(prefix.into());
         stream
@@ -77,12 +111,50 @@ impl EthEip1559TxArgs {
             .append(&format_address(&self.to))
             .append(&format_bigint(&self.value)?)
             .append(&self.input)
-            .append_list(access_list)
+            .append_list(access_list);
+        Ok(stream)
+    }
+
+    /// Returns the signed RLP-encoded message for the EIP-1559 transaction
+    pub fn rlp_signed_message(&self) -> anyhow::Result<Vec<u8>> {
+        let mut stream = self.message_rlp_stream()?;
+        stream
             .append(&format_bigint(&self.v)?)
             .append(&format_bigint(&self.r)?)
             .append(&format_bigint(&self.s)?)
             .finalize_unbounded_list();
         Ok(stream.out().to_vec())
+    }
+
+    /// Returns the unsigned RLP-encoded message for the EIP-1559 transaction
+    pub fn rlp_unsigned_message(&self) -> anyhow::Result<Vec<u8>> {
+        let mut stream = self.message_rlp_stream()?;
+        stream.finalize_unbounded_list();
+        Ok(stream.out().to_vec())
+    }
+
+    /// Constructs a signed message using EIP-1559 transaction args
+    pub fn get_signed_message(
+        &self,
+        from: Address,
+        eth_chain_id: EthChainId,
+    ) -> anyhow::Result<SignedMessage> {
+        ensure!(self.chain_id != eth_chain_id, "Invalid chain id");
+        let method_info = get_filecoin_method_info(&self.to, &self.input)?;
+        let message = Message {
+            version: 0,
+            from,
+            to: method_info.to,
+            sequence: self.nonce,
+            value: self.value.clone().into(),
+            method_num: method_info.method,
+            params: method_info.params.into(),
+            gas_limit: self.gas_limit,
+            gas_fee_cap: self.max_fee_per_gas.clone().into(),
+            gas_premium: self.max_priority_fee_per_gas.clone().into(),
+        };
+        let signature = self.signature()?;
+        Ok(SignedMessage { message, signature })
     }
 }
 
@@ -145,5 +217,16 @@ mod tests {
         let args = create_eip1559_tx_args();
         let signature = Signature::new(SignatureType::Delegated, vec![0u8; EIP_1559_SIG_LEN - 1]);
         assert!(args.with_signature(&signature).is_err());
+    }
+
+    #[test]
+    fn test_signature() {
+        let args = create_eip1559_tx_args();
+        let signature = Signature::new(SignatureType::Delegated, vec![0u8; EIP_1559_SIG_LEN]);
+        args.clone().with_signature(&signature).unwrap();
+
+        let sig = args.signature().unwrap();
+        assert_eq!(sig, signature);
+        assert!(args.to_verifiable_signature(sig.bytes().to_vec()).is_ok());
     }
 }