+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 - [#4937](https://github.com/ChainSafe/forest/pull/4937) Added
   `forest-cli f3 manifest` CLI command.
 
+- [#4706](https://github.com/ChainSafe/forest/issues/4706) Add support for the
+  `Filecoin.EthSendRawTransaction` RPC method.
+
 ### Changed
 
 ### Removed
@@ -58,6 +61,9 @@ F3 by default and includes initial power table CID on calibration network.
 
 - [#4917](https://github.com/ChainSafe/forest/issues/4917) Support `dnsaddr` in
   the bootstrap list.
+
+- [#4939](https://github.com/ChainSafe/forest/issues/4939) Fix
+  `Filecoin.EthBlockNumber` RPC method return type to be an `EthUInt64`.
 
 ### Changed